Moved some repetitive code (of mine) into a function
