+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="org.fundacionparaguaya.advisorapp.activities.MainActivity">
+        tools:context="org.fundacionparaguaya.advisorapp.activities.DashActivity">
 
     <!--Added for testing-->
     <org.fundacionparaguaya.advisorapp.viewcomponents.DashboardTabBarView
